requirements.txt
Flask
firebase-admin
google-auth
google-auth-oauthlib
google-auth-httplib2
flask-cors
firebase-admin
python-dotenv